chore: There should be no warnings when building our UI test crates. Warnings can interfere with correct snapshotting of stderr output for intentionally failing cases.

+++ ui-tests/cbindgen/tests/cbindgen/rust/cases/const_generics_arrayvec/src/lib.rs
@@ -12,7 +12,7 @@ pub struct ArrayVec<T, const CAP: usize> {
 #[unsafe(no_mangle)]
 pub unsafe extern "C" fn push(v: *mut ArrayVec<*mut u8, 100>, elem: *mut u8) -> i32 {
     if (*v).len < 100 {
-        (*v).xs[(*v).len] = elem;
+        (*v).xs[(*v).len as usize] = elem;
         (*v).len += 1;
         1
     } else {
